-fixed bug while drawing sideways blue bar -removed warnings and linker errors

+++ ConsoleApplication1/ConsoleRenderer.cpp
@@ -2,7 +2,7 @@
 
 void ConsoleRenderer::setConsoleToCursor()
 {
-	COORD pos = { cursor.x, cursor.y };
+	COORD pos = { (short)cursor.x, (short)cursor.y };
 	SetConsoleCursorPosition(console, pos);
 }
